Update cli-script.mdx
On my version of Linux (PopOS based on Ubuntu 22.04) the installed version of `esptool` was installed as `esptool.py`. I don't know if that is the case in all versions of Linux or in all installation methods (I used pip to install it) so I've phrased it to expect both versions.